pyproject.toml
[project]
name = "athena-assignment"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "numpy>=1.26.4,<2.1.0",
    "pandas>=2.2.2",
    "scikit-learn",
    "matplotlib",
    "jupyter>=1.1.1",
    "notebook>=7.5.3",
    "ipython>=9.9.0",
    "python-dotenv",
    "statsmodels>=0.14.6",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "pyarrow>=15.0.0",
    "pydantic>=2.12.5",
    "pmdarima>=2.1.1",
    "lightgbm>=4.6.0",
    "optuna>=4.7.0",
    "torch",
    "torchvision",
    "torchaudio",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/athena_analyze", "src/utils"]

[tool.uv]
dev-dependencies = ["pytest", "ruff"]

# --- PyTorch GPU環境向けの専用ソース設定 ---

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
# extra == 'gpu' の時だけ、PyTorch公式のCUDA12.4対応版を取得する設定
torch = [
  { index = "pytorch-cu124", marker = "extra == 'gpu'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "extra == 'gpu'" },
]
torchaudio = [
  { index = "pytorch-cu124", marker = "extra == 'gpu'" },
]
